Upgrade to Svelte 5
