--- lib/Boundary.js
@@ -1,4 +1,0 @@
-import Boundary from "./Boundary.svelte"
-import { createBoundary } from "@crownframework/svelte-error-boundary"
-
-export default createBoundary(Boundary);
