+++ lib/Wrapper.svelte
@@ -1,10 +1,17 @@
 <script>
-    import Component from "../src/Component.svelte"
-    import Boundary from "./Boundary.js"
+  import Component from "../src/Component.svelte"
 </script>
 
-<Boundary>
-  <Component {...$$props}>
-    <slot/>
+<svelte:boundary>
+  <Component {...$$restProps}>
+    <slot />
   </Component>
-</Boundary>
+  {#snippet failed(error, reset)}
+    <div class="plugin-error">Component failed to render: {error?.message}</div>
+  {/snippet}
+</svelte:boundary>
+
+<style>
+  .plugin-error { color: var(--spectrum-global-color-red-500); font-size: 0.9em }
+  .plugin-error:empty { display: none }
+</style>
